refactor(ai): extract turn request builder orchestration

<!-- mux-attribution: model=openai:gpt-5.6-sol thinking=high costs=3.41 -->
